pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-openrouter-app"
version = "0.1.0"
description = "FastAPI RAG MVP with OpenRouter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn==0.35.0",
  "python-multipart==0.0.20",
  "jinja2==3.1.6",
  "pypdf==6.0.0",
  "python-docx==1.2.0",
  "httpx==0.28.1",
  "chromadb==1.0.15",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
